Attempt to debug calculating total
